Set exception logged only when it really is emitted
